+++ gamry_parser/__init__.py
@@ -4,4 +4,5 @@ from .chronoa import ChronoAmperometry
 from .cv import CyclicVoltammetry
 from .eispot import Impedance
 from .ocp import OpenCircuitPotential
+from .squarewave import SquareWaveVoltammetry
 from .vfp600 import VFP600